+++ src/Forside/Forside.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
-import Infopanel from '../Infopanel/Infopanel';
+import Infopanel from './Infopanel/Infopanel';
 import './Forside.less';
-import IAwebpanel from '../IAwebpanel/IAwebpanel';
-import LegemeldtSykefravarPanel from '../Legemeldtsykefravarpanel/LegemeldtSykefravarPanel';
+import IAwebpanel from './IAwebpanel/IAwebpanel';
+import LegemeldtSykefraværPanel from './Legemeldtsykefraværpanel/LegemeldtSykefraværPanel';
 
 const Forside: React.FunctionComponent = () => {
     return (
         <div className="forside">
             <Infopanel />
-            <LegemeldtSykefravarPanel />
+            <LegemeldtSykefraværPanel />
             <IAwebpanel />
         </div>
     );
