+++ src/Forside/IAwebpanel/IAwebpanel.less
@@ -1,27 +1,27 @@
-@import '../variables';
+@import '../../variables';
 
-.infopanel {
-    border: 2px solid @navBlaLighten60;
+.iawebpanel {
+    margin-top: 1.5rem;
     padding: 2rem 2rem;
     display: flex;
     flex-direction: column;
-    justify-content: space-around;
 
     &__overskrift {
-        margin-bottom: 1.5rem;
+        margin-bottom: 1rem;
     }
 
     &__illustrasjon {
         margin-bottom: 1rem;
     }
 
+
     @media (min-width: @screen-sm-min) {
         padding: 2rem 2rem;
-        flex-direction: row-reverse;
-        position: relative;
+        flex-direction: row;
 
         &__tekst-wrapper {
-            margin-right: 2rem;
+            margin-left: 2rem;
         }
+
     }
 }
