+++ src/Forside/Infopanel/Infopanel.tsx
