+++ src/Forside/Legemeldtsykefraværpanel/LegemeldtSykefraværPanel.less
@@ -1,4 +1,4 @@
-@import '../variables';
+@import '../../variables';
 
 .legemeldtsykefravarpanel {
   margin-top: 1.5rem;
