+++ src/Forside/Legemeldtsykefraværpanel/Lesmeriawebpanel/EkspanderbartInnhold/ekspanderbartInnhold.less
