+++ src/Forside/Legemeldtsykefraværpanel/Lesmeriawebpanel/InfoToggler/InfoToggler.tsx
