+++ src/Forside/Legemeldtsykefraværpanel/Lesmeriawebpanel/lesmeriawebpanel.less
@@ -1,4 +1,4 @@
-@import '../../variables.less';
+@import '../../../variables.less';
 
 .lesmeriawebpanel {
   display: flex;
