+++ src/Forside/Legemeldtsykefraværpanel/Sykefraværsprosentpanel/Sykefraværsprosentpanel.less
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 .sykefravarsprosentpanel {
   display: flex;
