+++ src/Forside/Legemeldtsykefraværpanel/Sykefraværsprosentpanel/Sykefraværsprosentpanel.test.ts
